requirements.txt
dash==2.18.2
dash-mantine-components==0.14.7
dash-ag-grid==31.3.1
plotly==5.24.1
pandas==2.2.3
openpyxl==3.1.5
anthropic==0.42.0
python-dotenv==1.0.1
